requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
auth0-python==4.5.0
blinker==1.7.0
boto3==1.34.4
botocore==1.34.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
datadog-api-client==2.12.0
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.8
exceptiongroup==1.2.0
fal==0.11.3
fastapi==0.99.1
filelock==3.13.1
Flask==3.0.0
frozenlist==1.4.1
grpc-interceptor==0.15.4
grpcio==1.60.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
importlib-metadata==6.11.0
isolate==0.12.3
isolate-proto==0.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonify==0.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.2
opencv-python==4.8.1.78
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.1.0
portalocker==2.8.2
protobuf==4.25.1
PyAudio==0.2.14
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
PyYAML==6.0.1
replicate==0.22.0
requests==2.31.0
rich==13.7.0
s3transfer==0.9.0
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.1
starlette==0.27.0
structlog==22.3.0
tblib==3.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
urllib3==1.26.18
virtualenv==20.25.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
